added a function that triggers check for job completion
it's an intermediary to gta_hs_process_completed_jobs. In my view it's necessary since the role of the latter function changes now that we classify the results for each phrase separately. There is no need any longer to estimate the probabilities for all phrases in the job upon completion. They are already there. That reduces the role of process_completed_job to creating the XLSX and emailing it out (or comunicating it to RICARDO eventually). See changes to that function in subsequent commits.