removed processing round matching from phrase_list section of the query.
"phrase list section" is that part of the query that generates the list of potential phrases to check. It's the first part of the query.

The other part is the "user processed section" that generates the list of phrases to remove from the phrase list because the user has already worked on them in the current round.